requirements.txt
# PIDS Comparative Framework - Core Dependencies
# Python 3.8-3.10

# Deep Learning Frameworks
torch==1.12.1
torchvision==0.13.1
dgl==1.0.0
torch-geometric==2.1.0
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-cluster==1.6.0

# Data Processing
numpy==1.23.5
pandas==1.5.3
scipy==1.10.1
scikit-learn==1.2.2

# Graph Processing
networkx==2.8.8
graphviz==0.20.1
python-louvain==0.16

# Utilities
tqdm==4.65.0
pyyaml==6.0
xxhash==3.2.0
joblib==1.2.0
psutil==5.9.5

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Logging & Monitoring
wandb==0.15.0
tensorboard==2.13.0

# Database (for DARPA TC datasets)
psycopg2-binary==2.9.6
sqlalchemy==2.0.15

# Web & API
requests==2.31.0
flask==2.3.2
fastapi==0.95.2
uvicorn==0.22.0

# Configuration
hydra-core==1.3.2
omegaconf==2.3.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0
pre-commit==3.3.2
